requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
binaryornot==0.4.4
bleach==6.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.12
click==8.1.7
coloredlogs==15.0.1
comm==0.1.3
cookiecutter==2.3.0
cryptography==41.0.4
Cython==3.0.2
dataclasses-json==0.6.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
emoji==2.8.0
executing==1.2.0
fastapi==0.99.1
fastjsonschema==2.16.3
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
humanize==4.8.0
idna==3.4
importlib-resources==6.1.0
ipykernel==6.23.0
ipython==8.13.2
ipython-genutils==0.2.0
ipyvue==1.9.2
ipyvuetify==1.8.10
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
langchain==0.0.300
langsmith==0.0.40
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
Mesa==2.1.1
mistune==2.0.5
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook_shim==0.2.3
numexpr==2.8.6
numpy==1.24.3
onnxruntime==1.16.0
openai==0.28.0
overrides==7.4.0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pipfreeze==2.0.3
pipreqs==0.4.13
platformdirs==3.5.1
posthog==3.0.2
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.3
Pygments==2.15.1
pymdown-extensions==10.1
PyPika==0.48.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.6.15
python-json-logger==2.0.7
python-magic==0.4.27
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
reacton==1.7.1
regex==2023.8.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rich-click==1.6.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
solara==1.19.0
soupsieve==2.4.1
SQLAlchemy==2.0.21
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
text-unidecode==1.3
tiktoken==0.5.1
tinycss2==1.2.1
tokenizers==0.14.0
tornado==6.3.1
tqdm==4.66.1
traitlets==5.9.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.10.16
uri-template==1.2.0
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
widgetsnbextension==4.0.8
yarg==0.1.9
yarl==1.9.2
